server/requirements.txt
flask==3.1.0
pyyaml==6.0.2
gunicorn==23.0.0
gevent==24.11.1
